Move .sources.model to .model
